--- index.md
@@ -1,1 +1,0 @@
-[code coverage](./code/coverage/index.html)
