+++ scripts/expectThrow.js
